Extract diagram generation to diagrams.ts; replace GitHub raw SVG fetch with versioned CDN
- Move generateJPDiagram and generateCNDiagram out of WritingCard.tsx into src/diagrams.ts
- Extract shared renderSvgBlob helper to remove duplicated blob URL create/revoke pattern
- Add isCJK helper and early return when no chars found
- Add HanziCharData interface for typed charData instead of untyped json()
- Switch JP kanji source from raw.githubusercontent.com/KanjiVG to jsDelivr
  serving @madcat/kanjivg (now in package.json so version is tracked)
- Simplify JP stroke selection: @madcat/kanjivg SVGs have paths at SVG root,
  no StrokePaths group needed
- Switch CN source from hanzi-writer-data@latest to pinned @2.0.1 via jsDelivr
  (package also added to package.json)
- Remove console.log in ankiConnect.ts
