+++ src/WritingCard.tsx
@@ -6,148 +6,11 @@ import { addWritingCardToAnki } from "./api/ankiConnect";
 import { AppSettings } from "./types/AppSettings";
 import { WritingCardData } from "./types/Cards";
 import { AnkiConnectResult } from "./types/AnkiConnect";
+import { generateJPDiagram, generateCNDiagram } from "./diagrams";
 
 interface WritingCardProps {
   language: "jp-JP" | "zh-CN";
   settings: AppSettings;
-}
-
-const STROKE_COLORS = [
-  "#e74c3c",
-  "#e67e22",
-  "#2ecc71",
-  "#3498db",
-  "#9b59b6",
-  "#e91e63",
-  "#00bcd4",
-  "#f39c12",
-];
-
-async function generateJPDiagram(word: string): Promise<string | null> {
-  const chars = [...word].filter((ch) => {
-    const cp = ch.codePointAt(0)!;
-    return (cp >= 0x4e00 && cp <= 0x9fff) ||
-           (cp >= 0x3400 && cp <= 0x4dbf) ||
-           (cp >= 0xf900 && cp <= 0xfaff);
-  });
-  const size = 109; // KanjiVG viewBox is 109x109
-  const canvas = document.createElement("canvas");
-  canvas.width = size * chars.length;
-  canvas.height = size;
-  const ctx = canvas.getContext("2d")!;
-  ctx.fillStyle = "white";
-  ctx.fillRect(0, 0, canvas.width, canvas.height);
-
-  for (let i = 0; i < chars.length; i++) {
-    const cp = chars[i].codePointAt(0)!.toString(16).padStart(5, "0");
-    const url = `https://raw.githubusercontent.com/KanjiVG/kanjivg/master/kanji/${cp}.svg`;
-    try {
-      const res = await fetch(url);
-      if (!res.ok) continue;
-      const svgText = await res.text();
-
-      // Color individual strokes differently for visual clarity
-      const parser = new DOMParser();
-      const svgDoc = parser.parseFromString(svgText, "image/svg+xml");
-      const strokeGroup = svgDoc.querySelector('[id*="StrokePaths"]');
-      if (strokeGroup) {
-        const paths = strokeGroup.querySelectorAll("path");
-        paths.forEach((path, pi) => {
-          path.setAttribute(
-            "style",
-            `stroke:${STROKE_COLORS[pi % STROKE_COLORS.length]};stroke-width:3;fill:none;stroke-linecap:round;stroke-linejoin:round;`
-          );
-        });
-      }
-
-      const coloredSvg = new XMLSerializer().serializeToString(svgDoc);
-      const blob = new Blob([coloredSvg], { type: "image/svg+xml" });
-      const blobUrl = URL.createObjectURL(blob);
-
-      await new Promise<void>((resolve) => {
-        const img = new Image();
-        img.onload = () => {
-          ctx.drawImage(img, i * size, 0, size, size);
-          URL.revokeObjectURL(blobUrl);
-          resolve();
-        };
-        img.onerror = () => {
-          URL.revokeObjectURL(blobUrl);
-          resolve();
-        };
-        img.src = blobUrl;
-      });
-    } catch {
-      continue;
-    }
-  }
-
-  return canvas.toDataURL("image/png").split(",")[1];
-}
-
-async function generateCNDiagram(word: string): Promise<string | null> {
-  const chars = [...word];
-  const charSize = 120;
-  const padding = 8;
-  const canvas = document.createElement("canvas");
-  canvas.width = charSize * chars.length;
-  canvas.height = charSize;
-  const ctx = canvas.getContext("2d")!;
-  ctx.fillStyle = "white";
-  ctx.fillRect(0, 0, canvas.width, canvas.height);
-
-  for (let ci = 0; ci < chars.length; ci++) {
-    try {
-      const res = await fetch(
-        `https://cdn.jsdelivr.net/npm/hanzi-writer-data@latest/${chars[ci]}.json`
-      );
-      if (!res.ok) continue;
-      const charData = await res.json();
-
-      const innerSize = charSize - padding * 2;
-      // hanzi-writer-data uses bottom-left origin; flip Y with transform on path group
-      const paths = (charData.strokes as string[])
-        .map(
-          (d, i) =>
-            `<path d="${d}" fill="${STROKE_COLORS[i % STROKE_COLORS.length]}"/>`
-        )
-        .join("");
-
-      // Place stroke order numbers at the start of each stroke's median.
-      // Medians are in stroke-space coords (y increases upward); convert to SVG space: svgY = 900 - strokeY.
-      // Text is placed outside the flipped <g> so it reads correctly.
-      const numbers = (charData.medians as number[][][])
-        .map((median, i) => {
-          if (!median || median.length === 0) return "";
-          const [mx, my] = median[0];
-          const svgY = 900 - my;
-          return `<text x="${mx}" y="${svgY}" font-size="80" fill="white" stroke="#222" stroke-width="18" paint-order="stroke" text-anchor="middle" dominant-baseline="middle" font-family="sans-serif" font-weight="bold">${i + 1}</text>`;
-        })
-        .join("");
-
-      const svgStr = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1024 1024" width="${innerSize}" height="${innerSize}"><g transform="scale(1,-1) translate(0,-900)">${paths}</g>${numbers}</svg>`;
-      const blob = new Blob([svgStr], { type: "image/svg+xml" });
-      const blobUrl = URL.createObjectURL(blob);
-
-      await new Promise<void>((resolve) => {
-        const img = new Image();
-        img.onload = () => {
-          ctx.drawImage(img, ci * charSize + padding, padding, innerSize, innerSize);
-          URL.revokeObjectURL(blobUrl);
-          resolve();
-        };
-        img.onerror = () => {
-          URL.revokeObjectURL(blobUrl);
-          resolve();
-        };
-        img.src = blobUrl;
-      });
-    } catch {
-      continue;
-    }
-  }
-
-  return canvas.toDataURL("image/png").split(",")[1];
 }
 
 function WritingCard({ language, settings }: WritingCardProps) {
